requirements.txt
fpdf==1.7.2
pillow==10.3.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
